requirements.txt
-e .
aws_cdk.aws_ec2
aws-cdk.core
aws_cdk.aws_iam
aws_cdk.aws_rds
aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_autoscaling

